+++ src/components/fetch/FetchStop.tsx
@@ -4,10 +4,10 @@ import { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 
 import { BookmarkButton } from "../../features/bookmarks/BookmarkButton";
-import { settingsSelectors } from "../../features/settings/settingsSlice";
 import { EtaBusWithID, LineStopEta } from "../../models/etaObjects";
 import { EtaPredictionXml } from "../../models/etaXml";
 import { store } from "../../store";
+import { settingsSelectors } from "../../store/settings/slice";
 import { fluentStyles } from "../../styles/fluent";
 import CountdownGroup from "../countdown/CountdownGroup";
 import { CountdownRow } from "../countdown/CountdownRow";
