+++ src/index.tsx
@@ -4,13 +4,13 @@ import { Provider } from "react-redux";
 import { RouterProvider } from "react-router";
 
 import { saveStopBookmarkState } from "./features/bookmarks/localstorage";
-import { saveSettingsState } from "./features/settings/localstorage";
 import "./i18n/i18n";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
 import { router } from "./routes/MainRouter";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 import { store } from "./store";
+import { saveSettingsState } from "./store/settings/state";
 import { FluentTheme } from "./styles/FluentTheme";
 
 const container = document.getElementById("root");
